update to c bridging code

+++ C/bridging/src/bridging.h
@@ -3,7 +3,7 @@
 #ifndef BRIDGING_H
 #define BRIDGING_H
 
-double *bridging(graph_t *G, int *edgelist, double *scores);
-double *bridging_MPI(graph_t *G, int *edgelist, double *scores);
+double *bridging(graph_t *G, long *edgelist, double *scores);
+double *bridging_MPI(graph_t *G, long *edgelist, double *scores);
 
 #endif
