+++ C/bridging/src/io.h
@@ -4,8 +4,8 @@
 #include <graph_defs.h>
 
 int get_lines(FILE *f);
-int *read_edgelist_from_file(FILE *f, long *nNodes, long m, char **rev);
-int read_graph_from_edgelist(graph_t* G, int *EL, long n, long m);
+long *read_edgelist_from_file(FILE *f, long *nNodes, long m, char **rev);
+int read_graph_from_edgelist(graph_t* G, long *EL, long n, long m);
 
 #endif
